+++ kibot/gs.py
@@ -165,6 +165,7 @@ class GS(object):
     global_pcb_finish = None
     global_pcb_material = None
     global_remove_solder_paste_for_dnp = None
+    global_remove_solder_mask_for_dnp = None
     global_remove_adhesive_for_dnp = None
     global_resources_dir = None
     global_restore_project = None